+++ src/shaders/stop_points.wgsl
@@ -1,6 +1,6 @@
 struct ShaderConfig {
-    width: f32,  // how many pixels wide the image is
-    height: f32, // how many pixels high the image is
+    width: u32,  // how many pixels wide the image is
+    height: u32, // how many pixels high the image is
     bbox_min_lat: f32,
     bbox_min_lon: f32,
     bbox_max_lat: f32,
@@ -12,9 +12,9 @@ struct ShaderConfig {
 }
 
 struct JFAConfig {
-    jfa_width: f32,
-    jfa_height: f32,
-    jump_size: f32,
+    jfa_width: u32,
+    jfa_height: u32,
+    jump_size: u32,
     meters_per_px_x: f32,
     meters_per_px_y: f32,
 }
@@ -76,8 +76,8 @@ fn vs_main(
     let v = (config.bbox_max_lat - stop.lat) / (config.bbox_max_lat - config.bbox_min_lat);
 
     // uv -> pixel
-    let x_px = u * config.width;
-    let y_px = v * config.height;
+    let x_px = u * f32(config.width);
+    let y_px = v * f32(config.height);
 
     // meters -> pixels (approx)
     let mpp = min(jfa_config.meters_per_px_x, jfa_config.meters_per_px_y);
@@ -87,8 +87,8 @@ fn vs_main(
     let px = vec2<f32>(x_px, y_px) + local * radius_px;
 
     // pixel -> NDC
-    let ndc_x = (px.x / config.width) * 2.0 - 1.0;
-    let ndc_y = 1.0 - (px.y / config.height) * 2.0;
+    let ndc_x = (px.x / f32(config.width)) * 2.0 - 1.0;
+    let ndc_y = 1.0 - (px.y / f32(config.height)) * 2.0;
 
     var out: VsOut;
     out.pos = vec4<f32>(ndc_x, ndc_y, 0.0, 1.0);
